+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.0.0-alpha.8 - 2023-10-12
+
 - SDK only sends the `$feature_flag_called` event once per flag ([#47](https://github.com/PostHog/posthog-android/pull/47))
 - Groups are cached in the disk ([#48](https://github.com/PostHog/posthog-android/pull/48))